# openapi/kit-merged.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Realtime Service",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/oidc/keys/{set}": {
      "get": {
        "operationId": "KeyService_GetJsonWebKeySet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/keyJSONWebKeySet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "KeyService"
        ]
      },
      "delete": {
        "operationId": "KeyService_DeleteJsonWebKeySet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "KeyService"
        ]
      },
      "post": {
        "operationId": "KeyService_CreateJsonWebKeySet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/keyJSONWebKeySet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "keys": {
                  "$ref": "#/definitions/keyJsonWebKeySetGeneratorRequest"
                }
              }
            }
          }
        ],
        "tags": [
          "KeyService"
        ]
      },
      "put": {
        "operationId": "KeyService_UpdateJsonWebKeySet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/keyJSONWebKeySet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "keys": {
                  "$ref": "#/definitions/keyJSONWebKeySet"
                }
              }
            }
          }
        ],
        "tags": [
          "KeyService"
        ]
      }
    },
    "/v1/oidc/keys/{set}/{kid}": {
      "get": {
        "operationId": "KeyService_GetJsonWebKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/keyJSONWebKeySet"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "kid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "KeyService"
        ]
      },
      "delete": {
        "operationId": "KeyService_DeleteJsonWebKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "kid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "KeyService"
        ]
      },
      "put": {
        "operationId": "KeyService_UpdateJsonWebKey",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/keyJSONWebKey"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "set",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "kid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "key": {
                  "$ref": "#/definitions/keyJSONWebKey"
                }
              }
            }
          }
        ],
        "tags": [
          "KeyService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "keyJSONWebKey": {
      "type": "object",
      "properties": {
        "alg": {
          "type": "string",
          "required": [
            "alg"
          ]
        },
        "crv": {
          "type": "string",
          "x-nullable": true
        },
        "d": {
          "type": "string",
          "x-nullable": true
        },
        "dp": {
          "type": "string",
          "x-nullable": true
        },
        "dq": {
          "type": "string",
          "x-nullable": true
        },
        "e": {
          "type": "string",
          "x-nullable": true
        },
        "k": {
          "type": "string",
          "x-nullable": true
        },
        "kid": {
          "type": "string",
          "required": [
            "kid"
          ]
        },
        "kty": {
          "type": "string",
          "required": [
            "kty"
          ]
        },
        "n": {
          "type": "string",
          "x-nullable": true
        },
        "p": {
          "type": "string",
          "x-nullable": true
        },
        "q": {
          "type": "string",
          "x-nullable": true
        },
        "qi": {
          "type": "string",
          "x-nullable": true
        },
        "use": {
          "type": "string",
          "required": [
            "use"
          ]
        },
        "x": {
          "type": "string",
          "x-nullable": true
        },
        "x5c": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "y": {
          "type": "string",
          "x-nullable": true
        }
      },
      "required": [
        "alg",
        "kid",
        "kty",
        "use"
      ]
    },
    "keyJSONWebKeySet": {
      "type": "object",
      "properties": {
        "keys": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/keyJSONWebKey"
          }
        }
      }
    },
    "keyJsonWebKeySetGeneratorRequest": {
      "type": "object",
      "properties": {
        "alg": {
          "type": "string",
          "required": [
            "alg"
          ]
        },
        "kid": {
          "type": "string",
          "required": [
            "kid"
          ]
        },
        "use": {
          "type": "string",
          "required": [
            "use"
          ]
        }
      },
      "required": [
        "alg",
        "kid",
        "use"
      ]
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "description": "Authentication token, prefixed by Bearer: Bearer \u003ctoken\u003e",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}